Modernize Flix tooling
