+++ .github/workflows/build.yml
@@ -4,28 +4,78 @@ on:
   pull_request:
     branches: [ main ]
 
+permissions:
+  contents: read
+
 jobs:
   build:
     runs-on: ubuntu-latest
+    env:
+      FLIX_VERSION: v0.73.0
+      FLIX_SHA256: 5f69225d2e2a5c029abb5a89b306c97e2fe0b2dcb8f51316df9e7ec986811943
     steps:
-    - uses: actions/checkout@v3
-    - name: Set up JDK 11
-      uses: actions/setup-java@v3
+    - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
+    - name: Set up JDK 21
+      uses: actions/setup-java@ad2b38190b15e4d6bdf0c97fb4fca8412226d287 # v5.3.0
       with:
-        java-version: '11'
-        distribution: 'adopt'
-    - uses: actions/setup-node@v3
+        java-version: '21'
+        distribution: 'temurin'
+    - uses: pnpm/action-setup@0ebf47130e4866e96fce0953f49152a61190b271 # v6.0.9
       with:
-        node-version: '16'
-        cache: 'npm'
-    - run: |
-        curl -L -O https://github.com/flix/flix/releases/download/v0.27.0/flix.jar
-        java -jar flix.jar test
+        version: 11.8.0
+    - uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
+      with:
+        node-version: '24'
+        cache: 'pnpm'
+    - name: Install dependencies
+      run: pnpm install --frozen-lockfile
+    - name: Download Flix
+      run: |
+        curl --fail --location --proto '=https' --tlsv1.2 \
+          --output flix.jar \
+          "https://github.com/flix/flix/releases/download/${FLIX_VERSION}/flix.jar"
+        echo "${FLIX_SHA256}  flix.jar" | sha256sum --check -
+    - name: Test Flix package
+      run: java -jar flix.jar test
+    - name: Build Flix package
+      run: |
         java -jar flix.jar build
         java -jar flix.jar build-pkg
-        npm ci
-        npm run compute-hash
-        npx semantic-release
-        ./gradlew build
+        pnpm run compute-hash
+    - uses: actions/upload-artifact@043fb46d1a93c77aae656e7c1c64a875d1fc6a0a # v7.0.1
+      with:
+        name: release-assets
+        path: |
+          flix-semver2.fpkg
+          flix-semver2.fpkg.sha256
+          flix-semver2.fpkg.sha512
+        if-no-files-found: error
+
+  release:
+    if: github.event_name == 'push' && github.ref == 'refs/heads/main'
+    needs: build
+    runs-on: ubuntu-latest
+    permissions:
+      contents: write
+      issues: write
+      pull-requests: write
+    steps:
+    - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
+      with:
+        fetch-depth: 0
+    - uses: pnpm/action-setup@0ebf47130e4866e96fce0953f49152a61190b271 # v6.0.9
+      with:
+        version: 11.8.0
+    - uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
+      with:
+        node-version: '24'
+        cache: 'pnpm'
+    - name: Install dependencies
+      run: pnpm install --frozen-lockfile
+    - uses: actions/download-artifact@3e5f45b2cfb9172054b4087a40e8e0b5a5461e7c # v8.0.1
+      with:
+        name: release-assets
+    - name: Release
+      run: pnpm exec semantic-release
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
